fix: 修复删除文章分类接口问题

+++ src/main/java/com/springboot/springbootquickstart/controller/CategoryController.java
@@ -66,6 +66,10 @@ public class CategoryController {
   }
   @DeleteMapping("/delete")
   public Result delete(Integer id) {
+    Category category = categoryService.findCategoryById(id);
+    if (category == null) {
+      return Result.error("分类不存在");
+    }
     categoryService.delete(id);
     return Result.success();
   }
